chore(decouple-tick-snapshot): workspace-green verification + fix residual clippy (box 8.1)
Post-merge-burst integration gate: after the 9-PR merge (item-pickups,
platform-mechanics, liquid, overhead-map, mml, scenario-tdd, clippy-residual,
weapon, door, fullscreen), cargo test --workspace is fully green (no semantic
merge hazard), fmt clean, and 'cargo clippy -- -D warnings' (CI scope) clean.
Fixed the one residual --all-targets lint introduced by the merges
(e2e_tests.rs:620 redundant f32->f32 cast) so --all-targets is also clean.
Closes box 8.1.

+++ marathon-game/tests/e2e_tests.rs
@@ -617,7 +617,7 @@ mod marathon_game_level {
         match lights {
             LightData::Static(static_lights) => {
                 if let Some(light) = static_lights.get(idx) {
-                    let intensity = light.primary_active.intensity as f32 / 65536.0;
+                    let intensity = light.primary_active.intensity / 65536.0;
                     intensity.clamp(0.0, 1.0)
                 } else {
                     1.0
